Adding merge for logv4 storage

+++ src/main/java/com/belokur/jldbase/api/SegmentPosition.java
@@ -1,4 +1,4 @@
 package com.belokur.jldbase.api;
 
-public record SegmentPosition(Segment segment, long position) {
+public record SegmentPosition(Segment segment, int position) {
 }
